New configuration printer
The `--do-not-run` option has been added to allow inspection of a
configuration without running the given city.

By default, when an option is set in more than one place (multiple
places in the same config file, or in an included file aswell as an
including file), the printer shows the overridden values (the ones
which were specified before the last one). This can be switched off
with the `--no-overrides` option.

By default, the names of the files in which any value was
set (incuding overridden ones) is shown. By default, just the basename
of the file (everything after the last `/`) is shown. Whether and how
these file names are displayed can be controlled with the options

+ `--no-files`   : do not show the names at all
+ `--full-files` : show fully qualified file names

By default, the configuration is printed when any city is run. This
can be suppressed with `--hide-config`.