Add find_s12 (temporary fix).
currently function find_S12 returns an old-type objet (a class that
extends a dictionary with a pytables for persistency). Function
find_s12 (notice the difference between s12 and S12) returns simply a
dictionary that can then be passed to a high level S12 interface. The
old function is kept for the time being for backward compatibility.