Split HitCity into KrCity and HitCity
A KrCity computes and writes KrEvents, e.g, point-like events. This
means that the baricenter integrates all the time slices for a given
SiPM. Instead, a HitCity computes and writes HitEvents, in which the
baricenter (or any other position-finding algo) is computed
slice-by-slice. A HitCity is derived from a KrCity since both of them
use the same file_loop() function (which starts from pmaps).