Add functionality to KrCity.
Dorothea is now derived from KrCity, which centralizes the operations
of a city which computes and writes a KrEvent. In particular, in a
KrCity the position of the particle is computed integrating over all
temporal samples in SiPMs, but otherwise using the same
reco-algorithms (configurable) that can be used in HitCity (where
position is computed slice-by-slice).