Add daemons to IC
Daemons (see Nother Lights) are people's (external) souls in Lyra's
world. In IC daemons are slave (or companions) objects that can be
invoked from cities.

Consider for example the problem to add monitoring histograms to a
given city. One solution is to write all the monitoring code in the
city itself. Master would not like that for sure. The city becomes
immediately cumbersome, and the code booking and filling histograms
proliferates quickly. Plus, monitoring is tipically not stable, since
the critical histograms tend to vary with time.

The alternative is that the city calls a monitoring script (ms). At init
level, the city will call the init level of the ms (say to book
histograms). Then, at run level the city calls the run method of the
ms (fill histograms) and finally, at at the end level the city calls
the end method of the ms (save histograms to disk).

The scheme is quite general, and applies to any type of script that a
given city wants to invoke to delegate any task. Thus, a daemon is
defined as an object with an __init__() method, a run method and an
end method.

Daemons live in the new daemons directory (package) in IC. In this
commit, we define the base class and two toy daemons (Lyra and
Asriel). The daemons are defined by the variable deamons in the
corresponding city, which is associated to a list of names, each name
representing one daemon (the city will invoke all of them). The code
for daemon invocation is also defined in this commit.