Change pos to posxy in Cluster.
The refactoring of Cluster-Hit-Voxel has shown a feature that was
identified long ago and then forgotten. Namely, in a cluster, the
field called pos should be called posxy (since it describes a xy
point). This is now necessary to avoid conflict with the pos property
which is (ast it should) xyz.