Pass run_number to FEE and friends.
Diomira simulates the response of the PMTs FEE. It uses the data base
to pass the various calibration constants needed, including
deconvolution constants. However, it was assuming that run_number for
MC was *always* zero, as per the previous convention. This has changed
now (MC run number is negative and will match the corresponding run
number in data) and thus it is needed to propagate run_number all the
way up to FEE.