Protection agains empty events.
The Monte Carlo simulation produces occasional empty events, which are
characterized by flat waveforms running within the noise line. Irene
had already implemented a protection against those events, which
required that the sum of the zero-supressed waveforms were not
zero. However, a closely related cousin is an event whose
zero-supressed waveform yields a small number (due to noise
fluctuations) but still yields no s1 or s2 indexes on the search (e.g,
once a cut on 1-2 pes is imposed the waveform happens to be empty).

This commit expands the defence against such type of events, organizes
it in a function and adds counters for classification of the zombies.