Add new classes representing Tracks and Blobs.
Classes VoxelCollection, Track, Blob and TrackCollection have been
added.

A VoxelCollection is simply a collection of voxles.
A Track is a subclass of VoxelCollection defined by two features:
a collection of connected voxels and a pair of blobs,
defining its extremes.

A Blob is also a VoxelCollection, defining the set of Voxels in a
given radius near the extrema of the track.

A TrackCollection is a (writable) collection of Voxels.