Refactor reconstruction algorithms.
1.- The code no longer calls two algorithms (barycenter or corona)
which had to be selected by card, but a single algorithm (corona) with
a number of parameters. In particular, lm_radius can be set to a
negative number in which case corona will return an overall
barycenter (using all SiPms above Qthr threshold).

2. The condition imposing a single cluster when filling a kr_event has
been removed. That condition assumed that an overall barycenter
returning a single cluster *had to be used* in kr_city (e.g, to
compute KDSTs and maps). However, this is not a must and in fact,
better (less biased) results can in principle be obtained using a
corona algorithm which uses a limited number of SiPms around the local
maximum. On the other hand, the use of a corona may result in more
than one cluster in the tracking plane, even in the case of
single-particles (e.g, X-rays or Krypton), due, for example, to noisiy
SiPMs. However, the energy of the "true cluster" will be closer to the
energy of the S2 than the energy of any fake cluster. This criterium
is used to decide among clusters if there is more than one.

3. In the case of hit_city, if there is a reconstruction failure a hit
is still formed using a "NN cluster" (a Non-Numeric cluster). This
allows to keep track of the energy of the slice that failed to produce
a cluster.