Add function to read hits skiping NN and test.
Hits can be of type "NN" (No number), when the cluster algorithm fails
to find a XY position. Depending on the application the user may be
interested in skipping completely such hits. Thus, a function to
return only hits which are not of type NN has been added, plus the
corresponding test.