Add functions to load MC traks and hits.
A new module mchits_io containing functions
to read the MCTraks pytable and return event dictionaries is added.
Event dictionaries for MCTrks are provided by function read_mctracks
while event dictionaries for all the hits produced by and electron
or a positron in the event are produced by load_mchits.

A simple test that checks that the hits readout in one event
correspond to those in disk (hard wired in the code) is added.