Make Configuration.as_namespace read-only
The `as_namespace` property of the Configuration type provides a
convenient way of accessing the options set in an IC configuration,
but this namespace cannot be used to change the Configuration instance
from which its state was copied, as its contents are a copy of rather
than a reference to the original.

Any attempts to change this namespace, in order to set something in
the configuration, would fail silently.

Thus, rather than returning an `argparse.Namespace` from
`Configuration.as_namespace`, we use a subclass which raises a
`TypeError` with a helpful and explicit message when an attempt to
modify the namespace is made.