Fix voxelize_hits
The original algorithm calculated how many voxels were needed to cover
the whole bounding box of hits. This number was used to specify the
number of bins (voxels) into which the hits should be grouped, without
specifying the range of the bins. So `np.histogramdd` picked the range
of hit positions as the range of the bins. As this is very unlikely to
be an exact multiple of the desired voxel size, the voxels tended to
come out a bit smaller than required.

The solution is to explicity specify the spatial range which the bins
should cover, and ensure that it covers an exact multiple of voxel
sizes.