Make intervals half-open
This corresponds to a general definition of threshold and bin.

Threshold means that it is the minimum value accepted,
therefore >= is the right operator, particularly when dealing
with discrete values.

In the case of bins, we need a definition that does not
produce any misinterpretation. If we choose open intervals
and we define two ranges (a,b) and (b,c), the value b doesn't
fall in any of those, which doesn't feel right. On the other
hand if we choose close intervals and take [a,b] and [b,c],
there is ambiguity for the value b, which falls in both
cathegories. The right choice is a half-open interval. There
are, of course, two options, depending on which side you
want to be open. Following the reasoning for the threshold,
it seems natural to make the interval right-open rather than
left-open.

Besides, this fixes a mismatch between the python and the cython
implementations of the class minmax. This would have been
caught by the tests, provided they'd been there.