Clean up test_voxels_from_track_return_node_voxels
The purpose of the function is to act as a dead ringer for
`graph.nodes()`, providing a name which has a higher-level meaning in
IC.

The original test invoked a number of orthogonal IC functions. The new
version generates random graphs with hypothesis and checks that both
`test_voxels_from_track_return_node_voxels(graph)` and `graph.nodes()`
give identical output.