#384 Add size parameter to Voxel constructor
[author: jacg]
A number of situations are popping up where it would be useful to
query a voxel for its size, rather than having to hund backwards in
the code for the size that was requested, or having to infer the size
from the spacing.

[reviewer: paolafer]
The voxel size has been added to the class, in order to be able to
access it from the voxels themselves.