Major refactoring of the interpolation section in Correction class
The section of the class devoted to interpolation has been removed
following the principle of not reinventing the wheel. There is
an efficient and very handy function in scipy to do the job. As a
result of that, some extra posibilities arise, such as linear and
cubic interpolations. The original `RectBivariateSpline` interpolation
has been removed as there is no user case.