Treat both scalars and arrays nicely in Correction calls
This commit allows the class to be called with scalars, something we
already had originally, but that was lost during the refactoring. On
top of that, now the output is of the same type as the input, i.e. if
called with arrays, it returns an array, but if called with scalars,
it returns an scalar.