Revert "Refactor new_pmaps_io"
Although the version using groupby is visually more satisfying, it
introduces a mismatch between different sub-dataframes. For example:

df_A contains an event with peaks 0, 1, 2
df_B contains an event with peaks 0, 2

The previous version zips groupby objects, which would pair peak 0 in
df_A with peak 0 with df_B (ok) and peak 1 in df_A with peak 2 in
df_B. On top of introducing a mismatch between peaks, it would not run
over peak 2 in df_A.

The original version slicing dataframes manually is uglier (and
probably a bit slower for large dfs), but it is safe.