Improved sensor simulation
fee.NOISE_DAQ reduced to remove double counting
in noise stages of PMT electronics simulation.

fee.ADC_TO_PES recalculated using faked calibration
pulses and changed to 18.71 to better calculate the
gain dispersion of the PMTs.

In sensor_functions:

Gain dispersion application changed position so that
it doesn't affect the noise which was found not to match
dispersion in noise seen in data.

charge_fluctuation function added to simulate the single
pe fluctuation using the RMS values in the database.