445 Fix to potential error in NoiseSampler
https://github.com/nextic/IC/pull/445

[author: andLaing]

Code added baseline to samples before conversion to ADC which is an
error in the case of a baseline != 0.

Changed class to have access to sipm gains so that conversion could be
done in function.

Affected functions and tests altered to take change into account.

[reviewer: gonzaponte]

I would like to have a test that demonstrates the bug, but since it
depends on the database I reckon it will be too difficult to do
now. However, as soon as #420 is implemented, we should dedicate some
time to pay this debt.

I am happy with this PR otherwise. It introduces a bit of an overhead
as it includes another call to the database, which will be addressed
shortly.