Add test to ensure return values are correct
There is a subtle difference between the computation of the std
deviation in `numpy` and in `pandas`. The standard deviation is
defined as:

```
std = sum((x - x.mean())**2)/(N - ddof)
```

For `numpy`, ddof == 0 and for `pandas` ddof == 1, which makes a
subtle difference for large datasets and a moderate one for smaller
ones.

This ensures the output is exactly what we expect in a simple case.
As stated in the documentation of `np.std`:

> In standard statistical practice, ddof=1 provides an unbiased
> estimator of the variance of the infinite population. ddof=0
> provides a maximum likelihood estimate of the variance for normally
> distributed variables.

Thus, the correct version of `std` for our intentions is with ddof=1,
which is the default in `pandas`, but not in `numpy`.