480 Flatten mask in display_matrix
https://github.com/nextic/IC/pull/480

[author: jjgomezcadenas]

The display function display_matrix takes an optional mask array that
allows to plot only those values of the matrix accepted by the
mask. This feature had not been previously tested, and thus we had not
noted that the mask had to be flattened. This is now fixed.

[reviewer: gonzaponte]

Straightforward fix. Approved.