Remove commented code
The proposed behavior of the function when an exception of type
NoSuchNodeError is found is simply to print an error message, to
inform the user of the existence of an offending file and then ignore
the file. This is a case where it is justified not to abort the concat
of good files (which can take many minutes) because a spurious or
corrupted file is found. No need to throw a new exception either,
although it could be done if one wanted to let the user decide what to do.