Add work_in_python_version_no_test
By default we want the tests to run, to highlight pre-existing
problems when someone starts to work in/on IC. However, as this work
is leading up to automatic environment switching when historical
commits are checked out, we do not want to run all the tests on each
checkout. So we need a version that switches environment without
running the tests.