Stop failing to run parallel tests in work_in_python_version
The `manage.sh` command line argument sometimes specifies the python
version, sometimes the number of processors to use for parallel
testing. `work_in_python_version` accepts a python version, but
internally it calls run_tests_par, which requires the command line
argument to specify the number of processors. This conflict resulted
in the parallel tests not being run at all.

This is a quick hack to fix the problem as quickly as possible.