Tolerate warnings which don't concern us
`test_irene_split_trigger` used to check that exactly one
warning (with a very specific message) arose is a given
situation. When other, orthogonal, warnings appeared, the test failed.

The solution is to ensure that there is at least (rather than exactly)
one warning, and to check that the required message appears exactly
once in the collection of warnings that do appear.