515 Create new environment if necessary, when switching
[author: jacg]

This addresses a bug in #496: it (sometimes?) crashed when needing to
switch to an environment that had not been created yet.

The fix consits of ensuring that, if the environment does not yet
exist, it is created before the switch is attempted.

(I have no idea how to write automated tests for the stuff in
manage.sh, especially the ones related to switching environment
... which makes working on this very painful and error-prone.)

[reviewer: jjgomezcadenas]

The code is straight forward and can be approved right away. I share
the worry about automated tests here, but let's move this onwards.