Silence NaturalNameWarning
Hypothesis generates text which is used as a group name in a
pytable. This group is normally accessible as an attribute of the
table via Python's attribute access syntax. If the generated name
matches a Python keyword, such access would result in a syntax error,
even though `getattr(table, groupname)` would still work.

The NaturalNameWarning warns about this situation.

We silence it by rejecting names which match keywords.