Add `configure_liquid_test.py`
`configure_liquid_test.py` is just a copy of `configure_test.py` in which
the solid cities are being replaced with liquid ones.

This got through the net during the liquification, probably because
all other city tests were right next to the city implementation, while
these ones are tucked away somewhere else, out of sight.

As the liquid cities are so much simpler, the liquid version of the
dummy city is completely trivial: it just returns whatever arguments
it gets passed.

These tests use our sample config files, and these configs inject some
parameters which no longer exist (such as `write_mc_info` in
`penthesilea`).

The sample config files are hierarchical, and IIRC, we decided to get
rid of that. So I think that we need an new `liquid_config` directory
with new sample config files which conform to and document our new
config file structures. These can then coexist with the solid ones
until they are removed in my `goodbye-solid-cities` branch, which is
almost ready; the only thing that's missing is the liquified version
of these tests.

While doing this I also noticed that there are comments in the city
decorator which talk about simplifying the fiddling we do with the
configuration which arrives. We didn't deal with it at the time
because of the old structure of the config file. We should probably
open an issue reminding us to do this at some point. Probably not
right away.