578 Suppress health check for `test_pmap_event_id_selection`
https://github.com/nextic/IC/pull/578

[author: gonzaponte]

The test has been failing recurrently because of this known (and
harmless) issue. We don't know how to improve data generation at
present, so we suppress this specific health check.

[reviewer: jmbenlloch]

Until we find a solution to generate the samples in a faster way, this
is the only thing we can do to avoid infinite errors in travis.