580 Optimize mc_info_writer last_row parameter update
[author: mmkekic]

This PR changes the update of last_row parameter in mc_info_writer,
setting it to the value of the row number of the last written event,
which speeds up writing of sparse filtered events

[reviewer: msorel]

This PR optimises the way MC truth information is copied, avoiding to
start an event loop from the first MC event in the input file for
every MC event to be copied in the output. This particular event loop
now starts from the last written event, as it should.

In addition to passing all tests, it has been checked on a variety of
MC files, iincluding on events with randomly-ordered evt_numbers, that
MC information is identical before and after the fix.

In terms of performance, I could not find a performance improvement in
the particular application I tested (irene), but I have not seen a
deterioration either. It is quite possible that on different MC input
files and data processing streams a significant improvement can be
found.

To conclude, I would merge this as a potentially useful optimisation
for MC processing.