609 Eliminate not-used import
https://github.com/nextic/IC/pull/609

[author: paolafer]

I've eliminated a matplotlib import, which wasn't used, because it
caused a segmentation fault with a particular IC environment in the
majorana machine.

[reviewer: jmbenlloch]

Matplotlib sometimes can cause issues in headless machines, as has
happened to @paolafer while trying to run penthesilea. By removing
this import we avoid that situation in the future.