Fix how city decorator adds a default for detector_db
All IC cities needs a detector_db parameter to read the correct data
depending on the detector each job is using. But there are some
functions outside IC (in FANAL) that are also cities and does not
expect a detector_db parameter.

This parameter will have only a default value only if the decorated
function expect it.