Set of deconvolution used functions
Library with the different functions used in deconvolution:
- cut_and_redistribute_df basically cuts a dataframe on a given
condition and redistributes the values of the dataframe that are
passed.
- drop_isolated_sensors, as above, cuts a dataframe and redistributes
variables but based if a hit is considered isolated from the rest.
- deconvolutionInput prepares the data for deconvolution (puts it in a
grid and interpolates it if wanted, using interpolateSignal)
- find_nearest just finds the nearest array element to a given value,
useful to decide which PSF to use based on the Z position of the slice
- deconvolve basically applies the full deconvolution process,
including interpolation
- richardson_lucy is a modified version of the same function in
skimage, first the treatment of 0s has been changed as can lead to
nasty behaviour in the divisions, second the PSF mirroring has been
modified to make it n-dimensional, third an alternative stopping
criteria has been added to stop based on the relative change between
iterations and not on an absolute iteration number.