Beersheba city which applies track deconvolution
Beersheba reads esmeralda hits and applies deconvolution to
them. Afterwards, assign the cathode energy to the deconvolution
output and stores the deconvolved hits. Currently only tested for
slice by slice mode (2D deconvolution to each slice).

The PSF file that is used is a hdf5 containing a pandas dataframe
(GROUP: PSF, node: PSFs). Creation of the PSF shall go on another
commit not within IC but as a client, similar to KrCalib.

Features:
 - Signal can be interpolated or not before applying deconvolution.
 - Either uncorrected or corrected energy can be assigned.
 - Can use a one-step deconvolution (all EL and diffusion effects
 integrated within one PSF) or two-steps (first apply only-EL PSF
 deconvolution, then a gaussian-PSF deconvolution based on the drift
 of the hits).
 - Stopping criteria based on whatever is achieved first:
 maximum number of iterations or relative change between iterations threshold.