689 Fix coverage version
https://github.com/nextic/IC/pull/689

[author: jmbenlloch]

Travis has started to fail in IC installations due to a newer version
of the coverage package, which is not listed in our dependencies list
but it is need by one of them.

This PR simply fixes the coverage version manually so new
installations won't get the failing one.

[reviewer: paolafer]

This PR fixes a problem with hypothesis, caused by newer version of
the coverage package. The tests succeed, therefore I approve it.