Refactoring apply_deconvolution using pandas `apply` (rejected)
Although both the initial and the `apply` approach are, in nature, the
same, the `apply` approach has proven to be almost a factor two slower
than the initial implementation. This is due to the fact that apply is
called twice over the first group in ICs current version of
pandas (0.23).

This is solved in later versions so we could recover this
implementation once pandas is updated on IC.

cProfiles:
- Initial implementation:
https://github.com/Aretno/IC_profiles/blob/master/PR666_apply_deconvolution_initial_version.prof
- `apply` implementation (current pandas, 0.23):
https://github.com/Aretno/IC_profiles/blob/master/PR666_apply_deconvolution_commit_c90e327b_current_pandas.prof
- `apply` implementation (latest pandas, 1.0):
https://github.com/Aretno/IC_profiles/blob/master/PR666_apply_deconvolution_commit_c90e327b_latest_pandas.prof