Add test for Exception, Warning and Sum conservation.
create_waveform contains now an Exception and a Warning that must be
tested. Besides, the test cointains an assertion for Sum conservation:
that the number of initial pes is equal to the number of final counts,
if times values are restricted to the binning. Note this is not true
if times values are outside binning edges.