Modify 'load_dst' and 'load_dsts' to accept an event list.
Added argument to the usual load dst functions to just load events in
a given list. Usually loading several hdsts (and the alike) is a
memory heavy process as the full dst needs to be loaded and user does
not always need the full dst. This option allows for much faster and
lighter loading in such cases. If the new argument is not passed, the
function will behave as usual.

Use read_coordinates instead of looping through events (much faster)

Improve readability