Add blr function imitating the full procedure that happens in FPGA
A BLR function is added that computes the baseline as a moving
average. It applies the high-pass filter after substracting such
baseline (and thus it's done within the loop) and applies
deconvolution. As happens in the FPGA, the returned signal is equal to
the raw signal whenever the BLR is not active.