Add valdrada city
Adds valdrada, this city takes rwf and produces a dataframe with the
simulated trigger information. It takes a configurable set of
parameters that imitate the detector acquisition ones and their naming
convention. It provides all possible trigger candidates above a given
width and flags them as valid or not based on the configuration.