Change default compression to `None` in writers
This doesn't mean that the tables will not be compressed. Instead, the
compression configuration will fall back to a default value coming
from the overall file compression. However, if a non-None value is
passed, this will force that specific compression for the
corresponding nodes. This still allows for the (somewhat rare) cases
when an user wants to use these writers outside of a city.

From now on, *cities are responsible for setting the compression
configuration on file opening*