Add test file containing an event with an early low energy hit
This is a muon-induced xenon activation. The muon induces the recoil
of 137Xe which generates a very low energy hit. The 137Xe atom decays
much later producing the regular hits. The energy of the recoil hit
has been artificially reduced in order to have a more stable test.

The idea here is to have an event that "always" produces 0 electrons.