Use our miniconda distribution on GHA
For god knows how long we have asumed that there is no conda on GHA.
This means that GHA was not testing that our
run-this-command-and-everything-will-work setup program was working.

This fixes it. It involves an ugly hack in which we remove access to the
builtin conda and we setup our own version of conda in each step. To
avoid repetition, I have added this setup as a script in the repo.